7th pull
I've made some changes in switch.php and home.php because of the newest
files to turn on and off the leds. I've added new py scripts for turning
led on and off, and also for the sensors. I've changed the php file for
downloading the database backup to work in Raspbian - Linux